update embed example
