+++ packages/qvac-lib-infer-llamacpp-embed/examples/nativelog.js
@@ -1,10 +1,9 @@
 'use strict'
 
-const Corestore = require('corestore')
-const HyperDriveDL = require('@qvac/dl-hyperdrive')
+const FilesystemDL = require('@qvac/dl-filesystem')
 const GGMLBert = require('../index.js')
-const process = require('bare-process')
 const { setLogger, releaseLogger } = require('../addonLogging.js')
+const { downloadModel } = require('./utils')
 
 async function main () {
   console.log('Native Logger Example: C++ logging demonstration')
@@ -31,46 +30,31 @@ async function main () {
   console.log('Logger setup complete. C++ logging is now active.')
   console.log('Now demonstrating actual C++ logging during addon usage...\n')
 
-  // 2. Initializing data loader
-  const store = new Corestore('./store')
-  const hdStore = store.namespace('hd')
+  // 2. Downloading model
+  const [modelName, dirPath] = await downloadModel(
+    'https://huggingface.co/ChristianAzinn/gte-large-gguf/resolve/main/gte-large_fp16.gguf',
+    'gte-large_fp16.gguf'
+  )
 
-  const hdKey = 'd1896d9259692818df95bd2480e90c2d057688a4f7c9b1ae13ac7f5ee379d03e'
-  const hdDL = new HyperDriveDL({
-    key: `hd://${hdKey}`,
-    store: hdStore
-  })
+  // 3. Initializing data loader
+  const fsDL = new FilesystemDL({ dirPath })
 
-  // 3. Configuring model settings
+  // 4. Configuring model settings
   const args = {
-    loader: hdDL,
+    loader: fsDL,
     logger: console,
     opts: { stats: true },
-    diskPath: './models',
-    modelName: 'gte-large_fp16.gguf'
+    diskPath: dirPath,
+    modelName
   }
   const config = '-ngl\t25\nverbosity\t2'
 
-  // 4. Loading model
-  await hdDL.ready()
+  // 5. Loading model
   const model = new GGMLBert(args, config)
-  const closeLoader = true
-  let totalProgress = 0
-  const reportProgressCallback = (report) => {
-    if (typeof report === 'object' && Number(report.overallProgress) > totalProgress) {
-      process.stdout.write(
-        `\r${report.overallProgress}%: ${report.action} [${report.filesProcessed}/${report.totalFiles}] ${report.currentFileProgress}% ${report.currentFile}`
-      )
-      if (Number(report.currentFileProgress) === 100) {
-        process.stdout.write('\n')
-      }
-      totalProgress = Number(report.overallProgress)
-    }
-  }
-  await model.load(closeLoader, reportProgressCallback)
+  await model.load()
 
   try {
-    // 5. Generating embeddings
+    // 6. Generating embeddings
     const query = 'Hello, can you suggest a game I can play with my 1 year old daughter?'
     const response = await model.run(query)
     const embeddings = await response.await()
@@ -83,10 +67,9 @@ async function main () {
     console.error('Error occurred:', errorMessage)
     console.error('Error details:', error)
   } finally {
-    // 6. Cleaning up resources
+    // 7. Cleaning up resources
     await model.unload()
-    await hdDL.close()
-    await store.close()
+    await fsDL.close()
     releaseLogger()
   }
 }
